conductor: add support for automatic merge of LmP
When api/lmp/ is called with POST a script is called on the git
repository corresponding to the Projects defined in conductor. The
script will attempt to automatically merge lmp manfest into the
project's manifest and push the project manifest to the remote. Once
this happens (and there are any changes) the Foundries CI should build
the platform for the project which triggers automated testing.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@foundries.io>