core: allow different qa-reports name than project name
Sometimes it's not possible to test all required option in a single
factory. Results should still be reported in the same place in
qa-reports. For this reason allowing different name to be used for
calling qa-reports is required. This patch allows to use a separate
qa-reports project name, different than conductor project name.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@foundries.io>