core: Use lmp-manifest sha1 as a build version for qa-reports
When submitting test jobs for qa-reports SHA1 from lmp-manifest will be
used. The value is taken from the merge commit if it's on the HEAD of
the tree. Otherwise SHA1 of HEAD is used. This will finally allow to use
test results from different factories in a single qa-reports project.
The assumption is they will use the same version of upstream lmp-manifest.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@foundries.io>