scripts: automatically fix merge conflict with public keys
When conf/keys directory in public lmp-manifest is changed, this
produces a conflict on the merge to the factory lmp-manifest. This patch
addresses the issue by re-releting the public keys.

NOTE: With conductor project set to test on merge only, the testing
round will not be created when this happens. It's a rare event, so it
should not be a big issue in the long run.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@foundries.io>